Cleanup spec_helper; move code to classes in /spec/support dir
