+++ Rakefile
@@ -35,7 +35,7 @@ end
 task :default => :spec
 
 $LOAD_PATH.unshift(File.dirname(__FILE__))
-require 'spec/obfuscation'
+require 'spec/support/obfuscation'
 desc 'Obfuscates account.yml file.'
 task :obfuscate do
   Spec::Obfuscation.encrypt_file(File.join(File.dirname(__FILE__), 'spec', 'account.yml'))
